refactors router out of app into its own component
